fix: the position of judgment
